Move some ITCases to expression test and Merge Class to Accelerate Testing
